Version for demo at group meeting.
Added version of cyclops with partitioned timing recovery.
Reverted version of benchmarking due to a performance regression.
This makes this demo incompatible with MacOS at the moment.
It actually may be time to get rid of the dependecy on the benchmark
code since all of the telemety is handled internally at this point.